binfmt: rename rpecoff.h -> rpe.h (and matching .c / test)
Brings the PE / COFF spec header in line with the rest of binfmt's
naming convention: ELF uses relf.h + relfparser.h, Mach-O uses
rmacho.h + rmachoparser.h, but the third pair was named with the
asymmetric rpecoff.h + rpeparser.h. Tests inside test/rpecoff.c
were already registered as RTEST (rpe, ...) - the filename was the
only place still carrying the old name.
